requirements.txt
pandas==1.5.1
gspread==5.7.2
streamlit==1.22.0
oauth2client==4.1.3
python-dotenv==1.0.0
